merger/requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
Deprecated==1.2.13
idna==3.2
pycparser==2.20
PyGithub==1.55
PyJWT==2.2.0
PyNaCl==1.4.0
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wrapt==1.13.1
